feat: 점수 계산 객체 생성
