+++ src/test/java/chess/domain/BoardTest.java
@@ -113,14 +113,4 @@ public class BoardTest {
         Assertions.assertThat(board.findMovablePositions(startPosition))
                 .isEqualTo(new Path(List.of(upPosition, doubleUpPosition)));
     }
-
-    @Test
-    void 점수를_계산할_수_있다() {
-        Assertions.assertThat(board.calculateScore(Side.WHITE)).isEqualTo(38);
-    }
-
-    @Test
-    void 승리자를_계산할_수_있다() {
-        Assertions.assertThat(board.calculateWinner()).isEqualTo(Side.NEUTRALITY);
-    }
 }
